🚧 Extract pages types
