+++ src/pages/Registration/index.tsx
@@ -29,19 +29,14 @@ import { Theme, Values } from '../../constants';
 import HeaderLayout from '../../layouts/HeaderLayout';
 import FileOperations from '../../utils/FileOperations';
 
+// Type imports
+import { IDialogState } from '../../types/globals/Dialog';
+import { picturePath } from '../../types/pages/Registration';
+
 // Theme imports.
 import { styles, styledComponents } from './styles';
 
-// Type declarations.
-type picturePath = string | null;
-
 // Interface declarations.
-interface IDialogState {
-  open: boolean,
-  title: string,
-  message: string,
-}
-
 interface ISignUpForm {
   fullName: string,
   birthDate: Date | null,
