+++ src/services/chat/api.ts
@@ -1,7 +1,7 @@
 import firestore from '@react-native-firebase/firestore';
 import {
   DocumentData, DocumentRefData, Query, QuerySnapshot,
-} from '../../types/firebase';
+} from '../../types/services/firebase';
 
 const chatDocument = (chatUID: string): DocumentRefData => firestore().collection('chats').doc(chatUID);
