+++ src/types/globals/Carousel/index.ts
