+++ src/types/globals/Dialog/index.ts
@@ -1,0 +1,5 @@
+export interface IDialogState {
+  open: boolean,
+  title: string,
+  message: string,
+}
