+++ src/types/pages/Animal/index.ts
@@ -1,0 +1,21 @@
+import { DocumentRefData } from '../../services/firebase';
+
+export interface InterestedUser {
+  id: string;
+  ref: DocumentRefData;
+  imageURI: string;
+  userName: string;
+  birthDate: number;
+}
+
+export interface UserCircleProps {
+  user: InterestedUser;
+  callback: (user: InterestedUser) => void;
+}
+
+// Interface declarations.
+export interface IUploadedPicture {
+  id: string,
+  remoteName: string,
+  localUri: string,
+}
