+++ src/types/pages/Login/index.ts
@@ -1,0 +1,4 @@
+export interface ILoginForm {
+  email: string,
+  password: string,
+}
