+++ src/types/pages/Registration/index.ts
@@ -1,0 +1,1 @@
+export type picturePath = string | null;
